+++ InformationRetrievalManager/InformationRetrievalManager.NLP/InvertedIndex/IInvertedIndex.cs
@@ -14,7 +14,7 @@ namespace InformationRetrievalManager.NLP
         /// <param name="term">The term</param>
         /// <param name="documentId">Document ID</param>
         /// <exception cref="ArgumentNullException">If term is null or empty or document ID is not positive number.</exception>
-        void Put(string term, int documentId);
+        void Put(string term, long documentId);
 
         /// <summary>
         ///     <para>
