+++ InformationRetrievalManager/InformationRetrievalManager.NLP/Query/IQueryModel.cs
@@ -14,7 +14,7 @@ namespace InformationRetrievalManager.NLP
         /// </summary>
         /// <param name="data">Documents data(<see cref="InvertedIndex._vocabulary"/>)</param>
         /// <exception cref="ArgumentNullException">Missing data reference.</exception>
-        void CalculateData(IReadOnlyDictionary<string, IReadOnlyDictionary<int, IReadOnlyTermInfo>> data);
+        void CalculateData(IReadOnlyDictionary<string, IReadOnlyDictionary<long, IReadOnlyTermInfo>> data);
 
         /// <summary>
         /// Calculates model for query
@@ -24,7 +24,7 @@ namespace InformationRetrievalManager.NLP
         /// <param name="processingConfiguration">Processing configuration to use for the query index processing</param>
         /// <exception cref="ArgumentNullException">Missing query reference.</exception>
         /// <exception cref="InvalidOperationException">Missing parameters. <see cref="CalculateData"/> must be called beforehand.</exception>
-        void CalculateQuery(string query, IReadOnlyDictionary<string, IReadOnlyDictionary<int, IReadOnlyTermInfo>> data, IndexProcessingConfiguration processingConfiguration);
+        void CalculateQuery(string query, IReadOnlyDictionary<string, IReadOnlyDictionary<long, IReadOnlyTermInfo>> data, IndexProcessingConfiguration processingConfiguration);
 
         /// <summary>
         /// Calculate best matching documents by the query.
@@ -32,6 +32,6 @@ namespace InformationRetrievalManager.NLP
         /// <param name="select">Limit number of records to select (0 to ignore limit).</param>
         /// <returns>Sorted array of document IDs from the best matching to the least.</returns>
         /// <exception cref="InvalidOperationException">Missing parameters. <see cref="CalculateData"/> and <see cref="CalculateQuery"/> must be called beforehand.</exception>
-        int[] CalculateBestMatch(int select);
+        long[] CalculateBestMatch(int select);
     }
 }
